Add hololens holographic remoting in uwp PC
Add support of remouting of content to holo. Problem was in publishing setting. There is a newtwork check box in unity build settings. File-Buildsettings-Playersettings then in the inspector window under Publishing Settings there is Capabilities and a bunch of check boxes. Make sure internetClient and all of that is enabled.
Still have problem with kernel32.dll and msvcrt.dll